Bump StateCodelab to Alpha04
Changes:
- Migrate from Stack to Box
- Change imports for Box
- Restructure code for slot to use Box(modifier) { slot() } to workaround compiler bug
- Also extracted ExpandingCard.kt standalone UDF example to a separate file.
