+++ StateCodelab/start/build.gradle
@@ -42,7 +42,8 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
     buildFeatures {
-        compose true
+        // TODO: Re-enable when alpha04 and agp support adding compose via buildFeature
+//        compose true
         viewBinding true
     }
     composeOptions {
@@ -52,11 +53,14 @@ android {
 }
 
 dependencies {
+    // TODO: Remove this when alpha04 and agp support adding compose via buildFeature
+
+    kotlinCompilerPlugin "androidx.compose.compiler:compiler:$compose_version"
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.1'
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.2.0'
+    implementation 'com.google.android.material:material:1.2.1'
     implementation "androidx.fragment:fragment-ktx:1.2.5"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.2.0"
 
@@ -72,8 +76,8 @@ dependencies {
 
     testImplementation 'junit:junit:4.13'
     testImplementation "com.google.truth:truth:1.0.1"
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 }
 
 // Compiler flag to use experimental Compose APIs
